06/05 Second commit
I like to go through the github version of my work cause it looks so different from the atom version. So I made some stylistic changes and I'm now ready for the code review. I am yet to go back and change the color calculator because I just wanted to keep moving but I will go back and correct it.